Move GithubRepoAlg to VCSRepoAlg
